+++ src/css.ts
@@ -1,7 +1,14 @@
+declare const _RawCssString__Brand: unique symbol;
+/**
+ * Raw, un-transpiled CSS code.
+ * May be incomplete and/or include (SCSS-like) nested syntax
+ */
+export type RawCssString = string & { [_RawCssString__Brand]?: true };
+
 export const css = function (
   strings: TemplateStringsArray,
   ...values: Array<unknown>
-): string {
+): RawCssString {
   const valCount = values.length;
   return strings
     .map((str, i) => {